third_party/blink/web_tests/http/tests/preload/warning/request-method-expected.txt

CONSOLE WARNING: A preload for 'http://127.0.0.1:8000/preload/warning/resources/hello.txt' is found, but is not used because the request HTTP method does not match.
This is a testharness.js-based test.
Harness: the test ran to completion.
